faet: 안건 생성 + 회원 기능 추가 + 홈화면 갱신

+++ app/src/main/java/com/hgh/na_o_man/domain/model/agenda/AgendaRequestModel.kt
@@ -3,5 +3,5 @@ package com.hgh.na_o_man.domain.model.agenda
 data class AgendaRequestModel (
     val shareGroupId: Long,
     val title: String,
-    val photoIdList: List<Int>
+    val photoIdList: List<Long>
 )
